requirements.txt
redis
#git+git://github.com/aio-libs/aioredis.git@ea948d0#egg=aioredis
aioredis
uvloop==0.11.2
pyyaml
